server/requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
configobj==5.0.6
cryptography==3.4.7
dnspython==1.16.0
Flask==1.1.2
Flask-Cors==3.0.10
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pymongo==3.11.3
pyOpenSSL==20.0.1
requests==2.25.1
semantic-version==2.8.5
setuptools-rust==0.12.1
six==1.15.0
systematic==4.8.7
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
